+++ src/nustattools/stats/_derate.py
@@ -11,6 +11,8 @@ from numpy.typing import ArrayLike, NDArray
 from scipy.linalg import block_diag, sqrtm
 from scipy.stats import chi2
 
+from .gx2.functions import gx2inv
+
 
 @njit()  # type: ignore[misc]
 def _fix(cov: NDArray[Any]) -> NDArray[Any]:
@@ -141,9 +143,16 @@ def make_positive_definite(cov: NDArray[Any]) -> NDArray[Any]:
 
 
 def get_whitening_transform(
-    cov: NDArray[Any], transform: str = "mahalanobis"
+    cov: NDArray[Any],
+    transform: str = "zca_aligned",
+    projection: NDArray[Any] | None = None,
 ) -> tuple[NDArray[Any], NDArray[Any]]:
     """Get the blockwise whitening matrix and inverse."""
+
+    tokens = transform.split("_")
+    if len(tokens) > 2:
+        msg = f"Unknown whitening transform '{transform}'."
+        raise ValueError(msg)
 
     blocks = get_blocks(cov)
     W_l = []
@@ -158,21 +167,56 @@ def get_whitening_transform(
 
         c = make_positive_definite(c)
         # Determine transformation
-        if transform == "mahalanobis":
+        if tokens[0] in ("mahalanobis", "zca"):
             sc = sqrtm(c)
-            W_l.append(np.linalg.inv(sc))
-            Wi_l.append(sc)
-        elif transform == "cholesky":
+            W = np.linalg.inv(sc)
+            Wi = sc
+            W_l.append(W)
+            Wi_l.append(Wi)
+        elif tokens[0] in ("zca-cor",):
+            V = np.sqrt(np.diag(c))
+            Vi = 1 / V
+            V = np.diag(V)
+            Vi = np.diag(Vi)
+            cor = Vi @ c @ Vi
+            sc = sqrtm(cor)
+            W_l.append(np.linalg.inv(sc) @ Vi)
+            Wi_l.append(V @ sc)
+        elif tokens[0] in ("cholesky",):
             W = np.linalg.cholesky(np.linalg.inv(c)).T
             W_l.append(W)
             Wi_l.append(np.linalg.inv(W))
         else:
-            m = f"Unknown whitening transform '{transform}'."
-            raise ValueError(m)
+            msg = f"Unknown whitening transform '{transform}'."
+            raise ValueError(msg)
 
         i += n
 
-    return np.asarray(block_diag(*W_l)), np.asarray(block_diag(*Wi_l))
+    W = block_diag(*W_l)
+    Wi = block_diag(*Wi_l)
+
+    Rt_l = []
+    if len(tokens) == 2:
+        if tokens[1] == "aligned":
+            # Align whitened coordinates with image space
+            P = W @ projection @ Wi
+            i = 0
+            for n in blocks:
+                E = np.diag([0] * i + [1] * n + [0] * (len(P) - i - n))[:, i : i + n]
+                A = P[i : i + n, :] @ E
+                U, _, _ = np.linalg.svd(A)
+                Rt_l.append(U)
+
+                i += n
+
+            Rt = block_diag(*Rt_l)
+            Wi = Wi @ Rt
+            W = Rt.T @ W
+        else:
+            msg = f"Unknown whitening transform '{transform}'."
+            raise ValueError(msg)
+
+    return np.asarray(W), np.asarray(Wi)
 
 
 def derate_covariance(
@@ -180,9 +224,11 @@ def derate_covariance(
     *,
     jacobian: ArrayLike | None = None,
     sigma: float = 3.0,
-    precision: float = 0.01,
     return_dict: dict[str, Any] | None = None,
-    whitening: str = "mahalanobis",
+    whitening: str = "zca_aligned",
+    method: str = "gx2",
+    precision: float = 0.01,
+    max_batch_size: int = 10_000,
 ) -> float:
     """Derate the covariance of some data to account for unknown correlations.
 
@@ -193,27 +239,66 @@ def derate_covariance(
     cov : numpy.ndarray or list of numpy.ndarray
         The covariance matrix of the data or a list of covariances that add up
         to the total. Unknown covariance blocks must be ``np.nan``. Off
-        diagonal blocks may only be ``0'' or ``np.nan''. Diagonal blocks must
-        not be ``np.nan''.
+        diagonal blocks may only be ``0`` or ``np.nan``. Diagonal blocks must
+        not be ``np.nan``.
     jacobian : numpy.ndarray, default=None
         Jacobian matrix of the model prediction wrt the best-fit parameters.
     sigma : float, default=3.
         The desired confidence level up to which the derated covariance should
         be conservative, expressed in standard-normal standard deviations. E.g.
         ``sigma=3.`` corresponds to ``CL=0.997``.
-    precision : float, default=0.01
-        The derating factor is calculated using numerical sampling. This parameter
-        determines how many samples to throw. Lower values mean more samples.
     return_dict : dict, optional
         If specified, the nightmare covariance and thrown data samples are
         added to this dictionary for detailed studies outside the function.
-    whitening : ``"cholesky"`` or ``"mahalnobis"``, default="mahalanobis"
+    whitening : str, default="zca_aligned"
         Specify which method to use for the whitening transform.
+    method : str, default="gx2"
+        Either ``gx2`` or ``mc``. The former calculates the p-value directly,
+        the latter uses Monte Carlo methods.
+    precision : float, default=0.01
+        If the derating factor is calculated using numerical sampling (MC), this
+        parameter determines how many samples to throw. Lower values mean more
+        samples.
+    max_batch_size : int, default=10_000
+        If the derating factor is calculated using numerical sampling (MC), this
+        parameter determines how many samples to are thrown at once. This is
+        repeated until the total number for the required precision is reached.
 
     Returns
     -------
     a : float
         The derating factor for the total covariance.
+
+    Notes
+    -----
+
+    The basic available whitening transforms are:
+
+    ``mahalanobis`` or ``zca``
+        ``W = sqrtm(inv(cov))``
+
+    ``zca-cor``
+        ``W = sqrtm(inv(cor)) @ diag(1/sqrt(diag(cov)))``
+
+    ``cholesky``
+        ``W = cholesky(inv(cov)).T``
+
+    See [Kessy2015]_.
+
+    If ``_aligned`` is appended to a basic transform, an additional rotation
+    matrix is prepended, which aligns the whitened coordinate axes with the
+    model parameter space given by `jacobian`. See [Koch2024]_.
+
+    References
+    ----------
+
+    .. [Kessy2015] Kessy, Agnan / Lewin, Alex / Strimmer, Korbinian
+       "Optimal whitening and decorrelation",
+       The American Statistician 2018, Vol. 72, No. 4, pp. 309-314 , Vol. 72, No. 4,
+       Informa UK Limited, p. 309-314, https://arxiv.org/abs/1512.00809
+
+    .. [Koch2024] L. Koch "Hypothesis tests and model parameter estimation on
+       data sets1 with missing correlation information", TBD
 
     """
 
@@ -237,25 +322,33 @@ def derate_covariance(
     else:
         jacobian = np.asarray(jacobian)
 
+    # Projection matrix in original coordinates
+    S = make_positive_definite(cov_0)
+    Si = np.linalg.inv(S)
+    A = jacobian
+    Q = np.linalg.inv(make_positive_definite(A.T @ Si @ A)) @ A.T @ Si
+    P = A @ Q
+
+    # Target matrix
+    T = Si @ P
+
     # Transform to whitened coordinate systems and calculate "nightmare_cov"
     # covariance, then transform back
     nightmare_cov = np.zeros_like(cov_0)
     for c, c0 in zip(covl, cov_0_l):
-        # Determine the whitening transform for each covariance
-        W, Wi = get_whitening_transform(c, transform=whitening)
         # Whitened correlation is identity matrix
         cor = np.eye(len(c0))
         # Set unknowns back to NaN
         cor[np.isnan(c)] = np.nan
 
+        # Determine the whitening transform for each covariance
+        W, Wi = get_whitening_transform(c, transform=whitening, projection=P)
+
+        # Target matrix in whitened coordinates
+        Txi = Wi.T @ T @ Wi
+
         # Assumed total covariance in whitened coordinates
-        S = make_positive_definite(W @ cov_0 @ W.T)
-        Si = np.linalg.inv(S)
-        A = W @ jacobian
-        Q = np.linalg.inv(make_positive_definite(A.T @ Si @ A)) @ A.T @ Si
-        P = A @ Q
-        T = Si @ P
-        cor_nightmare = fill_max_correlation(cor, T)
+        cor_nightmare = fill_max_correlation(cor, Txi)
 
         # Transform back to non-whitened coordinates
         cov_nightmare = Wi @ cor_nightmare @ Wi.T
@@ -268,45 +361,104 @@ def derate_covariance(
     n_param = jacobian.shape[1]
     crit_0 = chi2.isf(alpha, df=n_param)
 
-    # Nightmare critical value from random throws
-    rng = np.random.default_rng()
-    # Matrix that solves the least squares problem
-    # Uses assumed covariance
-    parameter_estimator = (
-        np.linalg.inv(jacobian.T @ cov_0_inv @ jacobian) @ jacobian.T @ cov_0_inv
-    )
-    # Assumed covariance in parameter space
-    assumed_parameter_cov = parameter_estimator @ cov_0 @ parameter_estimator.T
-    assumed_parameter_cov_inv = np.linalg.inv(
-        make_positive_definite(assumed_parameter_cov)
-    )
-    # Actual nightmare_cov covariance
-    nightmare_parameter_cov = (
-        parameter_estimator @ nightmare_cov @ parameter_estimator.T
-    )
-    # Estimate necessary precision
-    # var = alpha(1-alpha) / (n f(crit_0)**2) =!= (crit_0 * rel_error)**2
-    n_throws = (
-        int(
-            (alpha * (1.0 - alpha))
-            / (chi2.pdf(crit_0, df=n_param) ** 2 * (crit_0 * precision) ** 2)
+    # Actual critical value in parameter space
+    crit_nightmare: np.floating[Any] | float = 0.0
+    if method == "gx2":
+        # Generalised chi-squared
+        Si_theta = A.T @ Si @ A
+        V_theta = Q @ nightmare_cov @ Q.T
+        H = Si_theta @ V_theta
+        weights = np.real(np.linalg.eigvals(H))
+        # Get rid of zeros and numerical artefacts
+        weights_l = sorted(weights[weights > 1e-2])
+        del weights
+        ndofs = list(np.ones_like(weights_l, dtype=int))
+        # Combine identical weights
+        i = 0
+        while i + 1 < len(weights_l):
+            if abs(weights_l[i] - weights_l[i + 1]) < 1e-3:
+                weights_l[i] = (
+                    weights_l[i] * ndofs[i] + weights_l[i + 1] * ndofs[i + 1]
+                ) / (ndofs[i] + ndofs[i + 1])
+                ndofs[i] += ndofs[i + 1]
+                del weights_l[i + 1]
+                del ndofs[i + 1]
+            else:
+                i += 1
+        tol = alpha / 1000.0
+        crit_nightmare = gx2inv(
+            alpha,
+            np.array(weights_l, dtype=float),
+            np.array(ndofs, dtype=int),
+            np.zeros_like(weights_l),
+            0,
+            0,
+            AbsTol=tol,
+            RelTol=tol,
+            side="upper",
         )
-        + 1
-    )
-    throws = rng.multivariate_normal(
-        mean=[0.0] * n_param, cov=nightmare_parameter_cov, size=n_throws
-    )
+    elif method == "mc":
+        # Nightmare critical value from random throws
+        rng = np.random.default_rng()
+        # Matrix that solves the least squares problem
+        # Uses assumed covariance
+        parameter_estimator = (
+            np.linalg.inv(jacobian.T @ cov_0_inv @ jacobian) @ jacobian.T @ cov_0_inv
+        )
+        # Assumed covariance in parameter space
+        assumed_parameter_cov = parameter_estimator @ cov_0 @ parameter_estimator.T
+        assumed_parameter_cov_inv = np.linalg.inv(
+            make_positive_definite(assumed_parameter_cov)
+        )
+        # Actual nightmare_cov covariance
+        nightmare_parameter_cov = (
+            parameter_estimator @ nightmare_cov @ parameter_estimator.T
+        )
+        # Estimate necessary precision
+        # var = alpha(1-alpha) / (n f(crit_0)**2) =!= (crit_0 * rel_error)**2
+        n_throws = (
+            int(
+                (alpha * (1.0 - alpha))
+                / (chi2.pdf(crit_0, df=n_param) ** 2 * (crit_0 * precision) ** 2)
+            )
+            + 1
+        )
+        # Throw in batches and average to avoid huge memory footprint
+        n_batches = (n_throws // max_batch_size) + 1
+        batch_size = (n_throws // n_batches) + 1
+        if batch_size * alpha < 2:
+            msg = f"Batch size of {batch_size} is too small for significance level {alpha}."
+            raise RuntimeError(msg)
+        throws = dist = None
+        for _ in range(n_batches):
+            del throws, dist
+            throws = rng.multivariate_normal(
+                mean=[0.0] * n_param, cov=nightmare_parameter_cov, size=batch_size
+            )
 
-    dist = np.einsum("ai,ij,aj->a", throws, assumed_parameter_cov_inv, throws)
-    crit_nightmare = -np.quantile(-dist, alpha)
+            dist = np.einsum("ai,ij,aj->a", throws, assumed_parameter_cov_inv, throws)
+
+            crit_nightmare += -np.quantile(-dist, alpha)
+
+        crit_nightmare /= n_batches
+    else:
+        msg = f"Unknown method {method}!"
+        raise ValueError(msg)
 
     derate = crit_nightmare / crit_0
+
+    if derate.ndim > 0:
+        derate = derate[0]
 
     derate = max(1.0, derate)
 
     if return_dict is not None:
         return_dict["nightmare_cov"] = nightmare_cov
-        return_dict["throws"] = throws
+        if method == "mc":
+            return_dict["throws"] = throws
+        return_dict["W"] = W
+        return_dict["Wi"] = Wi
+        return_dict["Q"] = Q
 
     return float(derate)